Fix format for column
