+++ pkg/collector/corechecks/snmp/internal/report/report_utils.go
@@ -10,6 +10,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/snmp/internal/checkconfig"
 	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/snmp/internal/valuestore"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
+	"strings"
 )
 
 func getScalarValueFromSymbol(values *valuestore.ResultValueStore, symbol checkconfig.SymbolConfig) (valuestore.ResultValue, error) {
@@ -70,4 +71,78 @@ func processValueUsingSymbolConfig(value valuestore.ResultValue, symbol checkcon
 		}
 	}
 	return value, nil
+}
+
+// GetTagsFromMetricTagConfigList retrieve tags using the metric config and values
+func GetTagsFromMetricTagConfigList(mtcl checkconfig.MetricTagConfigList, fullIndex string, values *valuestore.ResultValueStore) []string {
+	var rowTags []string
+	indexes := strings.Split(fullIndex, ".")
+	for _, metricTag := range mtcl {
+		// get tag using `index` field
+		if metricTag.Index > 0 {
+			index := metricTag.Index - 1 // `index` metric config is 1-based
+			if index >= uint(len(indexes)) {
+				log.Debugf("error getting tags. index `%d` not found in indexes `%v`", metricTag.Index, indexes)
+				continue
+			}
+			var tagValue string
+			if len(metricTag.Mapping) > 0 {
+				mappedValue, ok := metricTag.Mapping[indexes[index]]
+				if !ok {
+					log.Debugf("error getting tags. mapping for `%s` does not exist. mapping=`%v`, indexes=`%v`", indexes[index], metricTag.Mapping, indexes)
+					continue
+				}
+				tagValue = mappedValue
+			} else {
+				tagValue = indexes[index]
+			}
+			rowTags = append(rowTags, metricTag.Tag+":"+tagValue)
+		}
+		// get tag using another column value
+		if metricTag.Column.OID != "" {
+			// TODO: Support extract value see II-635
+			columnValues, err := getColumnValueFromSymbol(values, metricTag.Column)
+			if err != nil {
+				log.Debugf("error getting column value: %v", err)
+				continue
+			}
+
+			var newIndexes []string
+			if len(metricTag.IndexTransform) > 0 {
+				newIndexes = transformIndex(indexes, metricTag.IndexTransform)
+			} else {
+				newIndexes = indexes
+			}
+			newFullIndex := strings.Join(newIndexes, ".")
+
+			tagValue, ok := columnValues[newFullIndex]
+			if !ok {
+				log.Debugf("index not found for column value: tag=%v, index=%v", metricTag.Tag, newFullIndex)
+				continue
+			}
+			strValue, err := tagValue.ToString()
+			if err != nil {
+				log.Debugf("error converting tagValue (%#v) to string : %v", tagValue, err)
+				continue
+			}
+			rowTags = append(rowTags, metricTag.GetTags(strValue)...)
+		}
+	}
+	return rowTags
+}
+
+// transformIndex change a source index into a new index using a list of transform rules.
+// A transform rule has start/end fields, it is used to extract a subset of the source index.
+func transformIndex(indexes []string, transformRules []checkconfig.MetricIndexTransform) []string {
+	var newIndex []string
+
+	for _, rule := range transformRules {
+		start := rule.Start
+		end := rule.End + 1
+		if end > uint(len(indexes)) {
+			return nil
+		}
+		newIndex = append(newIndex, indexes[start:end]...)
+	}
+	return newIndex
 }
